refactor classes HomePage, LoginInputIncorrectDateTest, LoginInputCorrectDateTest, LoginPage, LoginPageXPath
