+++ src/main/java/ru/yandex/market/bazhenava/pages/home/DropDownMenuPageXPath.java
@@ -1,4 +1,4 @@
-package ru.yandex.market.bazhenava.pages;
+package ru.yandex.market.bazhenava.pages.home;
 
 public class DropDownMenuPageXPath {
